use u32 pointers for compression sources, they are supposed to be 4-byte aligned

+++ src/bios.rs
@@ -567,7 +567,7 @@ pub fn bit_unpack(src: *const u8, dest: *mut u32, params: *const BitUnpackParams
   }
 }
 
-pub fn lz77_uncomp_8bit(src: *const u8, dest: *mut u8) {
+pub fn lz77_uncomp_8bit(src: *const u32, dest: *mut u8) {
   #[cfg(not(all(target_vendor = "nintendo", target_env = "agb")))]
   {
     unimplemented!()
@@ -585,7 +585,7 @@ pub fn lz77_uncomp_8bit(src: *const u8, dest: *mut u8) {
   }
 }
 
-pub fn lz77_uncomp_16bit(src: *const u8, dest: *mut u16) {
+pub fn lz77_uncomp_16bit(src: *const u32, dest: *mut u16) {
   #[cfg(not(all(target_vendor = "nintendo", target_env = "agb")))]
   {
     unimplemented!()
@@ -603,7 +603,7 @@ pub fn lz77_uncomp_16bit(src: *const u8, dest: *mut u16) {
   }
 }
 
-pub fn huff_uncomp(src: *const u8, dest: *mut u32) {
+pub fn huff_uncomp(src: *const u32, dest: *mut u32) {
   #[cfg(not(all(target_vendor = "nintendo", target_env = "agb")))]
   {
     unimplemented!()
@@ -621,7 +621,7 @@ pub fn huff_uncomp(src: *const u8, dest: *mut u32) {
   }
 }
 
-pub fn rl_uncomp_8bit(src: *const u8, dest: *mut u8) {
+pub fn rl_uncomp_8bit(src: *const u32, dest: *mut u8) {
   #[cfg(not(all(target_vendor = "nintendo", target_env = "agb")))]
   {
     unimplemented!()
@@ -639,7 +639,7 @@ pub fn rl_uncomp_8bit(src: *const u8, dest: *mut u8) {
   }
 }
 
-pub fn rl_uncomp_16bit(src: *const u8, dest: *mut u16) {
+pub fn rl_uncomp_16bit(src: *const u32, dest: *mut u16) {
   #[cfg(not(all(target_vendor = "nintendo", target_env = "agb")))]
   {
     unimplemented!()
